mgnet/requirements.txt
numpy
opencv-python
torch
torchvision
pytorch_lightning
pyyaml
rich
tqdm
tensorboard
tensorboardX
matplotlib
